requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.1
redis==5.0.1
celery==5.3.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
phonenumbers==8.13.26
pillow==10.1.0
python-dotenv==1.0.0
python-slugify==8.0.1
